// types/builder.ts
export type TProduct = {
  id: number;
  name: string;
  type: string;
  description_short: string;
  description_long: string;
  code?: string;
  unit?: number | null;
  category?: number | null;
  cashback_type?: string;
  seo_title?: string;
  seo_description?: string;
  seo_keywords?: string[];
  global_category_id?: number | null;
  marketplace_price?: number;
  chatting_percent?: number | null;
  address?: string;
  latitude?: number | null;
  longitude?: number | null;
  photos?: string[] | null;
  manufacturer?: string | null;
  cashback_value?: number | null;
  external_id?: number | null;
  tags?: string[] | null;
  production_time_min_from?: string | null;
  production_time_min_to?: string | null;
  unit_name?: string | null;
  barcodes?: string[] | null;
  prices?: number[] | null;
  balances?: number[] | null;
  attributes?: string[] | null;
  group_id?: number[] | null;
  group_name?: string | null;
  is_main?: boolean | null;
  qr_hash?: string | null;
  qr_url?: string | null;
  updated_at?: number | null;
  created_at?: number | null;
};

export interface IProductActions {
  updateField: (key: keyof TProduct, val: any) => void;
  updatePhotos: (urls: string[]) => void;
  toggleField: (key: keyof TProduct) => void;
  addProduct: (name: string, price: number) => void;
  selectProduct: (id: number | string) => void;
  deleteProduct: (id: number | string) => void;
  clearAll: () => void;
  reset: () => void;
  generateAI: () => Promise<void>;
}